// mfes/workspace/src/utils/interfaces.ts
export interface UpdateCustomField {
 options: any;
 fieldId: string;
 value: string;
 type: string;
 label?: string;
 values?: string | string[];
 name?: string;
 isEditable?: boolean;
 order: number;
}